Create TwitterBot,YoutubeBot class, Tweet model
Create TwitterBot and YoutubeBot class to access @mention notifications for an authenticated user.
Check if the twitter user who tweets the @mention is a registered user in the db (Twitter gem).
Check if tweet's text contains '#radio' and a valid Youtube url (Youtube gem).
On pass of checks, create and insert new Tweet instance.